+++ core/reflection/advanced/advancedBinaryWriter.cpp
@@ -144,7 +144,7 @@ void AdvancedBinaryWriter::visit<std::string>(std::string &field, std::string /*
     if (stream == NULL) return;
     SYNC_PRINT(("AdvancedBinaryWriter::visit<std::string>(%s):called\n", field.c_str()));
     size_t pos = fieldOpening(fieldName);
-    SYNC_PRINT(("AdvancedBinaryWriter::visit<std::string> pos %d\n", stream->tellp()));
+    SYNC_PRINT(("AdvancedBinaryWriter::visit<std::string> pos %d\n", (int)stream->tellp()));
     writeString(stream, field);
     fieldClosing(pos);
